Ensure workflow_access will work with non-integer role identifiers.
In Backdrop, the unique identifier for a role is now the role name. This
requires some changes to how workflow_access manages node access. There
was some prior work in commit 3054dea42c for this but I've taken a
different approach and am instead storing a map in config between role
names and gids.

Fixes #7